roadwise/requirements.txt
Flask==0.12.2
google-cloud-datastore==1.4.0
gunicorn==19.7.1
Flask-SQLAlchemy==2.3.2
PyMySQL==0.8.0
Flask-PyMongo==0.5.1
PyMongo==3.6.0
six==1.11.0
google-api-python-client==1.7.4